PillarKataExercise - Roman Numerals Conversion
Consists of JUnit Testcases, to convert Roman strings to numerals and
numeric arabic values to Roman Strings.

Testcases are committed from the scratch, so multiple commits are done
to get used to the testing environment.

The conversion follows Test Driven developlment, which further lays the
scope for better understanding of the source code